requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.12
pytest==8.4.2
requests==2.32.5
python-dotenv==1.1.1
prometheus-client==0.21.1
python-json-logger==2.0.7
redis==6.4.0
celery==5.5.3
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-semantic-conventions==0.58b0
supabase==2.19.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.43
passlib[bcrypt]>=1.7.4
python-jose[cryptography]==3.3.0
solana==0.36.9
google-adk>=1.0.0
google-generativeai>=0.3.0
email-validator>=2.0.0
openai>=1.0.0
